feat: add validation in controller

+++ takeandfood/src/main/java/com/takeandfood/takeandfood/service/AnnouncementHandler.java
@@ -7,8 +7,10 @@ import com.takeandfood.takeandfood.beans.Announcement;
 import com.takeandfood.takeandfood.beans.Order;
 import com.takeandfood.takeandfood.dao.AnnouncementDao;
 import com.takeandfood.takeandfood.dao.OrderDao;
+import com.takeandfood.takeandfood.dao.PersonDao;
 import com.takeandfood.takeandfood.dto.AnnouncementDto;
 import com.takeandfood.takeandfood.mapper.AnnouncementMapper;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,20 +31,28 @@ public class AnnouncementHandler {
     private OrderDao orderDao;
 
     @Autowired
+    private PersonDao personDao;
+
+    @Autowired
     public AnnouncementHandler(AnnouncementDao announcementDao, AnnouncementMapper mapper) {
         this.announcementDao = announcementDao;
         this.mapper = mapper;
     }
 
     @Transactional
-    public void delete(Long id) {
-        orderDao.deleteWithAnnouncement(id);
-        announcementDao.delete(id);
+    public boolean delete(Long id) {
+        if(announcementDao.get(id).isPresent()) {
+            orderDao.deleteWithAnnouncement(id);
+            announcementDao.delete(id);
+            return true;
+        }
+        return false;
     }
 
     @Transactional
     public AnnouncementDto get(Long id) {
-        return mapper.toDto(announcementDao.get(id).orElse(null));
+        Optional<Announcement> announcement = announcementDao.get(id);
+        return announcement.map(value -> mapper.toDto(value)).orElse(null);
     }
 
     @Transactional
@@ -51,22 +62,33 @@ public class AnnouncementHandler {
 
     @Transactional
     public AnnouncementDto create(AnnouncementDto announcement) {
-        announcementDao.create(mapper.toEntity(announcement));
-        return announcement;
+        try {
+//            Announcement toCreate = ;
+            announcementDao.create(mapper.toEntity(announcement));
+            return announcement;
+        } catch(ObjectNotFoundException e) {
+            return null;
+        }
     }
 
     @Transactional
     public AnnouncementDto update(AnnouncementDto announcement) {
+        if(!announcementDao.get(announcement.getId()).isPresent()) {
+            return null;
+        }
         return mapper.toDto(announcementDao.update(mapper.toEntity(announcement)));
     }
 
     @Transactional
-    public List<AnnouncementDto> getOrderedByPerson(Long id) {
+    public List<AnnouncementDto> getOrderingsByPerson(Long id) {
+        if(!personDao.get(id).isPresent()) {
+            return null;
+        }
         List<Order> orders = orderDao.getForPerson(id);
         List<Announcement> announcements = new ArrayList<>();
         orders.forEach(e -> {
             Announcement announcement = announcementDao.get(e.getAnnouncement().getId()).orElse(null);
-            if(!Objects.isNull(announcement)) {
+            if (!Objects.isNull(announcement)) {
                 announcements.add(announcement);
             }
         });
