+++ takeandfood/src/main/java/com/takeandfood/takeandfood/service/FeedbackHandler.java
@@ -7,11 +7,13 @@ import com.takeandfood.takeandfood.beans.Feedback;
 import com.takeandfood.takeandfood.dao.FeedbackDao;
 import com.takeandfood.takeandfood.dto.FeedbackDto;
 import com.takeandfood.takeandfood.mapper.FeedbackMapper;
+import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -27,13 +29,18 @@ public class FeedbackHandler {
     }
 
     @Transactional
-    public void delete(Long id) {
-        feedbackDao.delete(id);
+    public boolean delete(Long id) {
+        if(feedbackDao.get(id).isPresent()) {
+            feedbackDao.delete(id);
+            return true;
+        }
+        return false;
     }
 
     @Transactional
     public FeedbackDto get(Long id) {
-        return mapper.toDto(feedbackDao.get(id).orElse(null));
+        Optional<Feedback> feedback = feedbackDao.get(id);
+        return feedback.map(value -> mapper.toDto(value)).orElse(null);
     }
 
     @Transactional
@@ -43,13 +50,20 @@ public class FeedbackHandler {
 
     @Transactional
     public FeedbackDto create(FeedbackDto feedback) {
-        Feedback feed = mapper.toEntity(feedback);
-        feedbackDao.create(feed);
-        return mapper.toDto(feed);
+        try {
+            Feedback toCreate = mapper.toEntity(feedback);
+            feedbackDao.create(toCreate);
+            return mapper.toDto(toCreate);
+        } catch(ObjectNotFoundException e) {
+            return null;
+        }
     }
 
     @Transactional
     public FeedbackDto update(FeedbackDto feedback) {
+        if(!feedbackDao.get(feedback.getId()).isPresent()) {
+            return null;
+        }
         return mapper.toDto(feedbackDao.update(mapper.toEntity(feedback)));
     }
 }
